tty/sbsauart: DO NOT UPSTREAM - make ttySBSA the active console device
The sbsauart driver doesn't register itself as a console
until module_initcall time. This allows the virtual console
driver to register the active console if no console= is
given on the cmdline. This patch allows ttySBSA to take
over the active console device role from any existing
console device if no console= is given on the cmdline.

This is just a temprary hack until SPCR table is supported.

Signed-off-by: Mark Salter <msalter@redhat.com>