irqchip / GIC: Add GIC version support in ACPI MADT
There is a field added in ACPI 6.0 MADT table to indicate the
GIC version, so parse the table to get its value for later use.

If GIC version presented in MADT is 0, we need to fallback to
hardware discovery to get the GIC version.

In ACPI MADT table there is no compatible strings to indicate
various irqchips and also ACPI doesn't support irqchips which
are not compatible with ARM GIC spec, so GIC version can be used
to load different GIC drivers which is needed for the later patch.

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>