irqchip / gic: Add stacked irqdomain support for ACPI based GICv2 init
As discussed in ACPI for ARM64 core patch set, we need to support
stacked irqdomains for ACPI based GIC init.

So introduce acpi_irq_domain for GIC core domain instead of referring
to the irq_default_domain, based on that, pass gsi as the argument and
get the gsi in gic_irq_domain_alloc() to add stacked irqdomain support
for ACPI based GICv2 init.

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>