irqchip / GICv3: remove the useless comparision of device node in xlate
In gic_irq_domain_xlate(), we match the domain's device node to the
controller and it turns out pretty useless, because we're always
registering the GIC domain with its device_node on DT, this is really
guaranteed to match.

Since we unify the way of matching irqdomain in DT and ACPI, this is
also a blocker of making this function usable in the context
of ACPI, so just remove it.

Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>