genirq: MSI: Add support to query MSI domain for a device
In OF, we normally use the msi-parent property for a device to associate
it to an MSI controller/domain. However, for ACPI, there are multiple ways
to associate this information.

For example, ACPIv5.1 can only support a single GIC MSI frame at the moment
since it doesn't have a way to associate a particular GIC MSI frame
to a device. ACPIv6 introduces IORT to define relationship b/w multiple ITS
and devices.

So, this patch introduces an interface for irqchip to register a callback,
to provide a way to determine appropriate MSI domain for a device.

Signed-off-by: Suravee Suthikulpanit <Suravee.Suthikulpanit@amd.com>